# internal/handler/api/api.swagger.yaml
openapi: 3.0.2
info:
  version: "1.0.0"
  title: Ticket booking
  description: Api for ticket booking
  contact:
    name: Ticket booking
    url: http://localhost
    email: asmaslik@mts.ru
servers:
  - url: http://localhost
tags:
  - name: routes
    description: маршруты самолета
  - name: passengers
    description: Бронирования для пассажиров

paths:
  /route:
    post:
      tags:
        - routes
      summary: добавить маршрут
      description: добавляет маршрут самолета
      operationId: addRoute
      requestBody:
        description: объект маршрута
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SRoute'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SIdentifier'
                  - $ref: '#/components/schemas/SRoute'

  /passengers/{route_id}:
    post:
      tags:
        - passengers
      summary: добавляем пассажира рейса
      description: добавляем пассажира рейса
      operationId: addPassenger
      parameters:
        - name: route_id
          in: path
          description: id маршрута самолета
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: объект пассажир
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SPassenger'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SIdentifier'
                  - $ref: '#/components/schemas/SPassenger'

components:
  schemas:

    SIdentifier:
      type: object
      properties:
        id:
          type: string
          format: uuid

    SRoute:
      type: object
      properties:
        route:
          type: string

    SPassenger:
      type: object
      properties:
        fullname:
          type: string
        age:
          type: integer
        phone:
          type: string
        email:
          type: string